backend/requirements.txt
# Production Dependencies for Render Hosting
# Core libraries for the DTF Color Extractor

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Image processing
Pillow>=9.0.0
opencv-python-headless>=4.8.0  # Headless version for server deployment
numpy>=1.24.0

# Background removal
rembg>=2.0.0

# File handling
python-multipart>=0.0.6

# Optional: For better performance
# opencv-contrib-python-headless>=4.8.0
